@@ -1,4 +1,4 @@
-import { Route, Switch } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 
 import AllMeetupsPage from "./pages/AllMeetups";
 import FavoritesPage from "./pages/Favorites";
@@ -7,17 +7,11 @@ import NewMeetupPage from "./pages/NewMeetup";
 function App() {
   return (
     <div>
-      <Switch>
-        <Route path="/" exact>
-          <AllMeetupsPage />
-        </Route>
-        <Route path="/new-meetup">
-          <NewMeetupPage />
-        </Route>
-        <Route path="/favorites">
-          <FavoritesPage />
-        </Route>
-      </Switch>
+      <Routes>
+        <Route path="/" exact element={<AllMeetupsPage />}/>
+        <Route path="/new-meetup" element={<NewMeetupPage />}/>
+        <Route path="/favorites" element={<FavoritesPage />}/>
+      </Routes>
     </div>
   );
 }
